Split api/index.tsx into seperate files & use typed imports
