+++ src/components/incident/SignedMessage.tsx
@@ -3,7 +3,10 @@ import React from "react";
 import ListItemText from "@material-ui/core/ListItemText";
 import Grid from "@material-ui/core/Grid";
 
-import { Timestamp } from "../../api";
+import classNames from "classnames";
+
+// Api
+import type { Timestamp } from "../../api/types.d";
 
 import { formatTimestamp } from "../../utils";
