+++ src/contexts.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useReducer } from "react";
 
-import { Filter } from "./api";
+import type { Filter } from "./api/types.d";
 import { filterReducer, FilterActions } from "./reducers/filter";
 import { initialUserState, userReducer, UserActions, UserStateType } from "./reducers/user";
 import { initialApiState, apiStateReducer, ApiStateActions, ApiState } from "./reducers/apistate";
