+++ src/reducers/filter.tsx
@@ -1,4 +1,4 @@
-import { Filter } from "../api";
+import type { Filter } from "../api/types.d";
 
 import { ActionMap, makeAction } from "./common";
